hygiene: -Wall -Wextra on core sources (2 sign-compares fixed), derive GC PERIOD, fix stale comments
Audit Step 4b. Standalone Release 40 / ASan 29 / TSan 29 green with ZERO core warnings; integration
rebuild rc=0, mode-2 verify-serve construct_BAD=0 (served=12990), mode-1 construct_BAD=0.

- Standalone CMakeLists now compiles accelerateMVCC.cpp / interval_list.cpp / epoch_table.cpp with
  -Wall -Wextra (vendored Kuku / gtest keep defaults). It surfaced exactly two benign sign-compares
  (no real bug -- the core is clean), now fixed: the ctor preload loop counter (int -> uint64_t) and
  can_pruning's active-id loop counter (int -> std::size_t). The build is now -Wall-clean so a future
  REAL sign-compare surfaces instead of hiding.
- accel_hook.cc: the GC driver's PERIOD is now DERIVED (EPOCH_SIZE * EPOCH_TABLE_SIZE / 4 via common.h)
  instead of a hardcoded 2500 that could silently diverge from the sizing constants.
- build_d1b3a.sh: a fresh integration build now compiles OUR sources with -Wall -Wextra (vendored Kuku
  stays -w), so the same warning signal reaches the integration build. (The live tree keeps -w until a
  reconfigure; the standalone already covers the shared core.)
- Corrected two stale/misleading accel_hook.cc comments (drainer "just counts", GC "empty body in step 1").

+++ include/accelerateMVCC.cpp
@@ -16,7 +16,7 @@ mvcc::Accelerate_mvcc::Accelerate_mvcc(uint64_t record_count, uint32_t kuku_log2
 
     this->epoch_table = new Epoch_table();
 
-    for (int i = 0; i < record_count; i++) {
+    for (uint64_t i = 0; i < record_count; i++) {
         kuku::item_type item = kuku::make_item(1, i);
 
         // value is header node pointer address for epoch-based interval linked list
